requirements.txt
playwright==1.45.0
langchain==0.2.5
langgraph==0.0.60
dspy-ai==2.4.0
openai==1.35.10
python-dotenv==1.0.1
pika==0.13.0
streamlit==1.36.0
streamlit-extras==0.4.0
python-telegram-bot==21.2
cryptography==42.0.8


